Review of report content and code
- Reviewed and corrected textual errors
- Checked code chunks for accuracy and
formatting
